Fix compilation on ARM with C++11: narrowing conversion
Cannot have double-to-float conversions inside { } in C++11.


cregit: 31f9df520527be3f7a0f45092dbb5bc9ed1700ff https://github.com/qt/qtbase/commit/31f9df520527be3f7a0f45092dbb5bc9ed1700ff

+++ src/gui/painting/qpainterpath_p.h
@@ -1156,7 +1156,7 @@ unit|}
 define|#
 directive|define
 name|KAPPA
-value|0.5522847498
+value|qreal(0.5522847498)
 end_define
 begin_endif
 unit|QT_END_NAMESPACE
